flask_server/requirements.txt
flask
flask_cors
pyarrow==14.0.0
python-dotenv
psutil
numpy
langchain
langchain_core
langchain_openai
langchain_qdrant
qdrant_client
langgraph
langgraph-checkpoint-sqlite
blinker
click
dnspython
Flask
Flask-Cors
itsdangerous
Jinja2
jmespath
MarkupSafe
pymongo
python-dateutil
python-dotenv
s3transfer
six
urllib3
Werkzeug